refactor the catalog discovery workflow
